category bar component is done. minor style changes.
