+++ Projects/Frontend/bazaar-frontend/src/screens/Home/home.js
@@ -9,6 +9,7 @@ import Col from "react-bootstrap/Col";
 import CardDeck from "react-bootstrap/CardDeck";
 import axios from "axios";
 import Card from "../../components/ProductCard/productCard";
+import CategoryBar from "../../components/category-bar/category-bar";
 import { serverUrl } from "../../utils/get-url";
 import { Link } from "react-router-dom";
 
@@ -32,78 +33,12 @@ class Home extends Component {
     axios.get(serverUrl + `api/product/`).then((res) => {
       this.setState({ products: res.data });
     });
-    axios.get(serverUrl + "api/product/categories/").then((res) => {
-      let resp = res.data;
-      let categoryStructureTemp = {};
-      let keys = [];
-      let categoryDictTemp = {};
-      for (let i = 0; i < resp.length; i++) {
-        if (resp[i]["parent"] == "Categories") {
-          keys.push(resp[i]["name"]);
-          categoryDictTemp[resp[i]["id"]] = resp[i]["name"];
-        }
-      }
-      this.setState({ categoryList: keys });
-      this.setState({ categoryDict: categoryDictTemp });
-      for (let i = 0; i < keys.length; i++) {
-        let sublist = [];
-        for (let j = 0; j < resp.length; j++) {
-          if (resp[j]["parent"] == keys[i]) {
-            sublist.push(resp[j]["name"]);
-          }
-        }
-        categoryStructureTemp[keys[i]] = sublist;
-      }
-      categoryStructureTemp[""] = [];
-      console.log("category strructure: " + categoryStructureTemp);
-
-      this.setState({ categoryStructure: categoryStructureTemp });
-    });
   }
 
   render() {
-    let active = 2;
-    let categoryList = this.state.categoryList;
-    let categoryStructure = this.state.categoryStructure;
-    let categories = [];
-    for (let number = 0; number < categoryList.length; number++) {
-      let subs = [];
-      let subList = categoryStructure[categoryList[number]];
-      if (subList) {
-        for (let subnumber = 0; subnumber < subList.length; subnumber++) {
-          subs.push(
-            <a
-              className="dropdown-item"
-              href={"/category/" + subList[subnumber]}
-            >
-              {subList[subnumber]}
-            </a>
-          );
-        }
-      }
 
-      categories.push(
-        <Pagination.Item
-          key={number}
-          className="myPaginationItem dropdown"
-          href={"/category/" + categoryList[number]}
-        >
-          <a
-            className="nav-link dropdown-toggle"
-            data-toggle="dropdown"
-            aria-haspopup="true"
-            aria-expanded="false"
-          >
-            <span className="mr-1"></span> {categoryList[number]}
-          </a>
-          <div className="dropdown-menu">{subs}</div>
-        </Pagination.Item>
-      );
-    }
-
-    let categoryDict = this.state.categoryDict;
+    
     let productCards = this.state.products.map((product) => {
-      console.log(JSON.stringify(product));
       return (
         <Col sm="3">
           <Card product={product}></Card>
@@ -115,9 +50,7 @@ class Home extends Component {
       <div>
         <div className="background">
           <Container>
-            <div className="myPagination">
-              <Pagination size="lg">{categories}</Pagination>
-            </div>
+           <CategoryBar></CategoryBar>
             <div className="row homeJumbotron">
               <Jumbotron style={{ background: "transparent" }}>
                 <h1>Welcome to Bazaar!</h1>
